+++ lib/mongoid.rb
@@ -145,3 +145,9 @@ module Mongoid #:nodoc
     ActiveModel::Observing::ClassMethods.public_instance_methods(false) <<
     { :to => Config })
 end
+
+# After loading Mongoid, if Rails is defined we will get the load hooks cranked
+# up here. This optimizes the number of calls to instantiate_observers.
+if defined?(Rails)
+  ActiveSupport.run_load_hooks(:mongoid, Mongoid)
+end